+++ NEWS.md
@@ -26,6 +26,7 @@
 23. Removed ```obsolete/``` directory.
 24. Added parentheses to installation script's additional info for options 1 and 2.
 25. Fixed ```rpse_broadcast_waitUntilInterval()``` and made it an non-static function.
+26. Added GPL3 info to ```ONEWS.md``` file.
 
 # Older NEWS information #
 Older news information can be found in the ```ONEWS.md``` archive file.